feat: Support regular expressions for S3 bucket includes/excludes
This commit updates the `aws-s3-include-buckets` and `aws-s3-exclude-buckets` configuration options to support regular expressions instead of wildcards. This change provides more flexibility and precision when specifying which S3 buckets should be included or excluded during data synchronization.

The README.md file has been updated to reflect this change in functionality and provide examples of how to use regular expressions for bucket filtering.